Use DataInteractiveState instead of uiState to manage global dragging state.
