+++ v3/src/components/drag-drop/attribute-drag-overlay.tsx
@@ -24,7 +24,7 @@ export function AttributeDragOverlay ({ activeDragId, overlayHeight, overlayWidt
   }
 
   // Drags initiated by plugins can specify the size of the overlay
-  const style: CSSProperties | undefined = overlayHeight || overlayWidth
+  const style: CSSProperties | undefined = overlayHeight && overlayWidth
     ? { height: `${overlayHeight}px`, width: `${overlayWidth}px` } : undefined
 
   // Drags initiated by plugins have to be offset based on the location of the plugin
