+++ v3/src/models/ui-state.ts
@@ -43,13 +43,6 @@ export class UIState {
   @observable
   private _interruptionCount = 0
 
-  @observable private _draggingDatasetId = ""
-  @observable private _draggingAttributeId = ""
-  @observable private _draggingXOffset = 0
-  @observable private _draggingYOffset = 0
-  @observable private _draggingOverlayHeight = 100
-  @observable private _draggingOverlayWidth = 100
-
   constructor() {
     makeObservable(this)
   }
@@ -76,30 +69,6 @@ export class UIState {
 
   get interruptionCount() {
     return this._interruptionCount
-  }
-
-  get draggingDatasetId() {
-    return this._draggingDatasetId
-  }
-
-  get draggingAttributeId() {
-    return this._draggingAttributeId
-  }
-
-  get draggingXOffset() {
-    return this._draggingXOffset
-  }
-
-  get draggingYOffset() {
-    return this._draggingYOffset
-  }
-
-  get draggingOverlayHeight() {
-    return this._draggingOverlayHeight
-  }
-
-  get draggingOverlayWidth() {
-    return this._draggingOverlayWidth
   }
 
   isFocusedTile(tileId?: string) {
@@ -152,39 +121,6 @@ export class UIState {
 
   @action setAttrIdToEdit(attrId?: string) {
     this.attrIdToEdit = attrId
-  }
-
-  @action setDraggingDatasetId(datasetId?: string) {
-    this._draggingDatasetId = datasetId ?? ""
-  }
-
-  @action setDraggingAttributeId(attributeId?: string) {
-    this._draggingAttributeId = attributeId ?? ""
-  }
-
-  @action setDraggingXOffset(offset = 0) {
-    this._draggingXOffset = offset
-  }
-
-  @action setDraggingYOffset(offset = 0) {
-    this._draggingYOffset = offset
-  }
-
-  @action setDraggingOverlayHeight(height = 100) {
-    this._draggingOverlayHeight = height
-  }
-
-  @action setDraggingOverlayWidth(width = 100) {
-    this._draggingOverlayWidth = width
-  }
-
-  @action endDrag() {
-    this.setDraggingAttributeId()
-    this.setDraggingDatasetId()
-    this.setDraggingOverlayHeight()
-    this.setDraggingOverlayWidth()
-    this.setDraggingXOffset()
-    this.setDraggingYOffset()
   }
 
   getRulerStateVisibility(key: RulerStateKey) {
